update tpke client api
