+++ .github/workflows/workspace.yml
@@ -77,6 +77,27 @@ jobs:
       - run: cargo check --all-features
       - run: cargo test --release --all-features
 
+  wasm-test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        rust:
+          - 1.63 # MSRV
+          - stable
+        target:
+          - wasm32-unknown-unknown
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions-rs/toolchain@v1
+        with:
+          profile: minimal
+          toolchain: ${{ matrix.rust }}
+          target: ${{ matrix.target }}
+          override: true
+      - run: cargo install wasm-pack
+      - run: wasm-pack test --node
+        working-directory: tpke-wasm
+
   benchmark:
     runs-on: ubuntu-latest
     needs: [ test ]
