+++ tpke/src/api.rs
@@ -4,21 +4,41 @@
 
 // TODO: Refactor this module to deduplicate shared code from tpke-wasm and tpke-wasm.
 
-use std::convert::TryInto;
-
 use ark_ec::{AffineCurve, ProjectiveCurve};
 use ark_ff::{BigInteger256, ToBytes};
+use std::convert::TryInto;
 
 // Fixing some of the types here on our target engine
 // TODO: Consider fixing on crate::api level instead of bindings level
-type E = ark_bls12_381::Bls12_381;
-type TpkePublicKey = ark_bls12_381::G1Affine;
-type TpkePrivateKey = ark_bls12_381::G2Affine;
-type TpkeCiphertext = crate::Ciphertext<E>;
-type TpkeDecryptionShare = crate::DecryptionShareFast<E>;
-type TpkePublicDecryptionContext = crate::PublicDecryptionContextFast<E>;
-type TpkeSharedSecret =
+pub type E = ark_bls12_381::Bls12_381;
+pub type TpkePublicKey = ark_bls12_381::G1Affine;
+pub type TpkePrivateKey = ark_bls12_381::G2Affine;
+pub type TpkeCiphertext = crate::Ciphertext<E>;
+pub type TpkeDecryptionShare = crate::DecryptionShareFast<E>;
+pub type TpkePublicDecryptionContext = crate::PublicDecryptionContextFast<E>;
+pub type TpkeSharedSecret =
     <ark_bls12_381::Bls12_381 as ark_ec::PairingEngine>::Fqk;
+pub type TpkeResult<T> = crate::Result<T>;
+
+pub fn encrypt(
+    message: &[u8],
+    aad: &[u8],
+    pubkey: &TpkePublicKey,
+) -> TpkeCiphertext {
+    let rng = &mut rand::thread_rng();
+    crate::encrypt(message, aad, pubkey, rng)
+}
+
+pub fn decrypt_with_shared_secret(
+    ciphertext: &TpkeCiphertext,
+    aad: &[u8],
+    shared_secret: &TpkeSharedSecret,
+) -> TpkeResult<Vec<u8>> {
+    crate::decrypt_with_shared_secret(ciphertext, aad, shared_secret)
+}
+
+// TODO: There is previous API implementation below. I'm not removing it to avoid breaking bindings.
+//  Review it and decide if we need it.
 
 #[derive(Clone, Debug)]
 pub struct PrivateDecryptionContext {
@@ -92,11 +112,14 @@ impl DecryptionShare {
     }
 }
 
+// TODO: Reconsider contents of ParticipantPayload payload after updating server API.
+
 #[derive(Clone, Debug)]
 pub struct ParticipantPayload {
     pub decryption_context: PrivateDecryptionContext,
     pub ciphertext: TpkeCiphertext,
 }
+
 impl ParticipantPayload {
     pub fn new(
         decryption_context: &PrivateDecryptionContext,
